requirements.txt
absl-py==2.0.0
addict==2.4.0
attrs==23.1.0
backcall==0.2.0
brotlipy==0.7.0
cachetools==5.3.2
click==8.1.7
codecov==2.1.13
colorama==0.4.6
coloredlogs==15.0.1
coverage==7.2.7
cycler==0.11.0
decorator==5.1.1
einops==0.6.1
exceptiongroup==1.2.0
flake8==3.9.2
flatbuffers==23.5.26
fonttools==4.38.0
ftfy==6.1.1
google-auth==2.25.2
google-auth-oauthlib==0.4.6
grpcio==1.60.0
humanfriendly==10.0
imageio==2.31.2
importlib-metadata==6.7.0
iniconfig==2.0.0
interrogate==1.5.0
ipdb==0.13.13
ipython==7.34.0
isort==4.3.21
jedi==0.19.1
kiwisolver==1.4.5
lmdb==1.4.1
Markdown==3.4.4
MarkupSafe==2.1.3
matplotlib==3.5.3
matplotlib-inline==0.1.6
mccabe==0.6.1
mkl-fft==1.3.1
mkl-service==2.4.0
mpmath==1.3.0
networkx==2.6.3
numpy==1.21.6
oauthlib==3.2.2
onnxruntime==1.14.1
opencv-python==4.8.1.78
packaging==23.2
parso==0.8.3
pexpect==4.9.0
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==4.0.0
pluggy==1.2.0
prompt-toolkit==3.0.43
protobuf==3.20.3
ptyprocess==0.7.0
py==1.11.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycodestyle==2.7.0
pyflakes==2.3.1
Pygments==2.17.2
pyparsing==3.1.1
pytest==7.4.3
pytest-runner==6.0.1
python-dateutil==2.8.2
pytorch-msssim==1.0.0
PyWavelets==1.3.0
PyYAML==6.0.1
regex==2023.10.3
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.19.3
scipy==1.7.3
sympy==1.10.1
tabulate==0.9.0
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tifffile==2021.11.2
toml==0.10.2
tomli==2.0.1
tqdm==4.66.1
traitlets==5.9.0
typing_extensions==4.7.1
wcwidth==0.2.12
Werkzeug==2.2.3
yapf==0.40.2
zipp==3.15.0
